requirements.txt
# Helix bot — pin major deps; upgrade intentionally.
anthropic>=0.40
python-telegram-bot>=21.0
python-dotenv>=1.0
pinecone>=5.0
openai>=1.0
